clear out old resume function and rctx
when we have two xlats in a row, the first one might need to be
resumed, so it pushes a resume ctx onto the stack.  It then returns
"done", and the xlat evaluator goes to the next xlat node.

However, the next xlat node might _not_ push a resume function onto
the stack.  So when the evaluator resumes, it will see the first
resume function and rctx, and call it again.  Cue boom.

The solution is on DONE, replace the resume function / rctx with
a NULL one which does nothing.