TMPL_TYPE_UNRESOLVED -> TMPL_TYPE_DATA_UNRESOLVED
just to make it clear that the tmpl contains something which will
be resolved to data, and not to an xlat, exec, attr, etc.

This is part of the enum cleanups, where we will allow the following
as enum names:

	::Foo
	'Foo'
	"Foo"

but not

	Foo

Which will eventually be interpreted as an attribute reference.