VSA types are decoded as groups
note that we will have to audit all of the "decode raw" routines.
the current behavior is for the _called_ function to create a
"raw" parent if the decode fails.  If the parent is itself a group,
the called function will instead have to return an error, and then
the calling function will turn it into a raw attribute