use correct escaping
now that we've fixed fr_asprint() in commit b10640, and the
escaping in 41193841eb, this test also needs fixing.  In order to
pass \' to a shell which takes a string in ' quotes, the config
file string must use one layer of escaping for FR, and a second
layer of escaping for the shell.

this issue highlights the fact that we likely need a "q" operator
ala Perl.