enforce parent / child relationships for structural pairs
so that if we have nested TLVs, the children have to go into the
parent, subject to some limitations.

This requirement helps to enforce correct structure.

Note that children can still go into "internal" group attributes,
which means that they can still go into the top-level request list.

This isn't ideal, but it's not clear how to fix that without also
breaking it for control and session-state