stop on empty buffer, too.
The fr_pair_list_afrom_str() function should really be converted
to sbuffs, but that isn't immediately trivial.

It expect to see NUL-terminated C strings, and will misbehave
if given bad data.  Happily, all of the stgring value-boxes in
the server are zero-terminated, so this isn't much of an issue.

The same goes for input from unit_test_attribute.