add test for large users file
the default buffer size is 8K, so we add a file larger than that.
If the sbuff extension only looks for one byte at a time, then
we will run out of data in the buffer while there is still pending
data to read.  The users file reader will then return an error.

By updating the sbuffs to read more than 1 character, we ensure
that there's enough data in the buffer for most parsing purposes.