document clang 16+ suppressions
-Wunsafe-buffer-usage is useless for C code, as it complains
about essentially every use of pointers.

-Wcast-function-type-strict complains about casting functions
which take (void*) to functions which take a typed pointer.  This
pattern is used extensively in the dlist wrappers.

-Wreserved-identifier complains about identifiers which begin
with "_".  This is used all over the place in our code, but is
generally forbidden in C.

-Wc2x-extensions complains about initializers like "foo = {}"
which is fine, but which isn't C99.

-Wextra-semi-stmt complains about ";;" at the end of lines.
Our code is clean, but some system header files are crap.