Avoid misleading errors in tmpl tokenizing
When resolving an internal attriubute, often a lookup will first have
been done in a protocol dictionary, which results in an error being
recorded.
However, if the attribute was subsequently found in the internal
dictionary, that error is meaningless - but will still be printed if
something else causes errors to print.