Attempt to keep coverity from being silly (CID #1551707)
For some unknown reason, Coverity skips the declaration of subst in
fr_vasprintf_internal()...and said declaration initializes it, so
that later, when substr is used, Coverity complains that subst is
not initialized!

To try to avoid this; we move the declaration of subst out of the
do {} while () loop it was in, to the outer block of the function
body, and then just assign to it in the loop.