correctly handle inter-dictionary dependencies.
when adding them, make the "next" one depend on the one which
loaded it.  Then add the "next" one to the autofree list for the
current dictionary.

When freeing the global context, walk through all of the
dictionaries and free the autoref references first. Then walk through
them again, freeing the externally loaded references.