tcp: fix set non-blocking on accept of connections
The socket to listen/accept for new connections was set as
non-blocking. However, upon accepting a new connection, the new
descriptor for it wasn't, causing the server to block on reads
indefinitely in some cases.

Fix this by calling fr_nonblock() on the new descriptor.
An alternative fix would be calling accept4() with SOCK_NONBLOCK.