Move to init/free for library, and not for each test case
now that we (mostly) have library-based init/free, we do not need
to do that for every test case.  Either the library is properly
initialized and it works, or the library isn't initialized, and
the tests fail.

Some protocols still need to be converted to use fr_dict_protocol_t

	ethernet, tacacs, tftp, vmps