clean up autofree
The only remaining errors are:

test.unit.dns, which fails to clean up the DNS dictionaries.
All of the other protocols are fine, so there's some extra magic
here.

test.keywords: fr_dict_free() is asked to remove a dependency
which doesn't exist.  That _should_ be an assertion, but for now
it's a recoverable error which lets other people get work done.