allow relative references in "clone="
remove code which allowed clone to reference other protocols.
The "reference protocol" code was changed during the refactoring
to support

	ATTRIBUTE ... group ref=..DHCPv4

which now makes this code invalid.  So we fix up this code to work,
and (at least temporarily) remove the ability to reference other
protocols.

Now that the framework supports relative references, it should be
fairly simply to re-add protocol references as a separate commit.