simplify and handle more corner cases
remove the difference between write_partial and write_cancelled.
it turns out they're both largely the same.

Which then simplifies the cancel() function, as it just sets
item->cancelled, which is much better.

ensure that if we're writing a partial packet, that the timers
are removed.  Because if writes are blocked, there's no reason
to run the retries to write more packets.

There's still an @todo to flush the outbound retry list when
write becomes unblocked.