Correct uninitialized scalar false positives (CID #1504037, #1503938)
Coverity doesn't realize that uninitialized char arrays used as
buffer space for sbuffs are in fact set by sbuff functions that
write to the sbuff. (Yes, we tried modeling, but it didn't work.)
The solution is to keep the sbuff FR_SBUFF_OUT() creates around
long enough to use fr_sbuff_start(&sbuff) "instead of" the array
(even though it really is the array); that satisfies coverity.