Add a `sent` flag to trunk requests
So that only the first time of sending a request counts towards
connection statistics.

Requests which are re-queued typically need to run on the same
connection (e.g. a series of SQL queries forming a transaction) and bad
things will happen if `max_uses` causes a connection to close part way
through such requests.