Add support for OpenLDAP and 389ds PBKDF2 passwords
This commit enables the PAP module to recognize and handle OpenLDAP
and 389ds style PBKDF2 passwords.
A new password type was added that leaves the password header intact
when passing it to the authentication module, because it contains
the hash function used. Currently, password_process_header strips
that out, as none of the other supported hash formats need it.