Python script to add an lldb command
This, once imported with an lldb "script" command, adds a "dd" command
to display certain types of values in an intelligible fashion using the
existing foo_debug() functions not only is it intelligible, it uses a
format people are used to.

The foo_debug() functions take a pointer to the structure. The dd command
will take names of variables that have the structure type as well, and
pass their address.