Do size calculation in appropriate type (CID #1604603)
In the error message in lst_expand(), the calculation of the size
has to be done in size_t. The cast of sizeof(void *) to uint32_t
causes it to be done in uint32_t, but that has the possibility of
overflow; note that n_capacity may be set to UINT_MAX, which would
guarantee it. The format conversion is changed to match.