Eliminate dead code (CID #1603583)
Since args was just initialized to NULL and never changed, the
xlat_func_args_set() call done if args is non-NULL never happens.
(Said call is the only use of xlat_redundant_args.) Coverity rightly
notices the dead code.

That said, it's hard not to think that the original author didn't
want to set args to a non-NULL value at some point. This PR will
solve the Coverity issue and won't change semantics, but at some
point it merits reconsideration.