Attempt Coverity-only overflow check in xlat_print() (CID #1604615)
The error return subtracts a size_t from an ssize_t, which C does
in size_t, and then casts the result to ssize_t for return, which
in theory raises the possibility of overflow. We add a Coverity-only
check that tries to catch it. If Coverity doesn't recognize it,
we'll have to fall back to annotation.