move fr_pair_print() to using ::enum-name
and update the regression tests to match.

Anything which reads the server output will need to be updated
to allow ::enum-name.

the value-box and tmpl functions already allow, but don't require
the "::" prefix for enum names.  So any server code will be OK.

i.e. radclient etc. will parse either ::enum-name or enum-name.

This change should only affect the unit tests, and the detail file
writer.