clean up uses of "set leaf num"
in most cases it's not necessary.  We instead initialize the
ar filter to NONE.

In other cases (now that we may have multiple kinds of filters),
we check that we're only setting the num for FILTER_INDEX.
and if we're clearing the num by setting it to NUM_UNSPEC, we
also set the filter to FILTER_NONE

Much of the code in the server assumed that it could always look
at ar->ar_num.  It would therefore behave "surprisingly" when
other filters are used.