don't cast xlat function arguments
if the output of the xlat is supposed to be cast, that shouldn't
affect the inputs to the xlat arguments.

In fact, we should probably update xlat_tokenize_argv() to cast
each argument as it's parsing, which should serve as a hint to
the parser as to what to expect.