check the right parent
the outer parent_map is taken from starting to parse the section.
map_afrom_cs() will parse multiple child sections, and call the
validation function for each one.  But will only pass the top-level
parent_map.

So fixup the ref to use the actual parent when there are multiple
levels