Remove level of annoying indirection and copying things from the fr_dict_protocol_t to the dict
From comments this seems mostly to avoid cleaning up all the references to default_type_size and subtype_table.

Load the protocol symbol directly using dlsym, as this'll run every time the library is referenced and not just the first time.

Allocate a dictionary attribute extension for protocol specific data if requested to by the fr_dict_protocol_t