Use common parsing functions for all references.
Add support for '@', so that '..' works as one would expect.

Stop clones adding a pre-cloned version of the attribute to the dictionary.

Allow clones, and enum references to reference foreign attributes, but only when they share the same base protocol.

Use the file and line numbers from refs, instead of from the fixup structs.