Change the default idle timeout to 3600 for incoming connections
The client is the one who initiates the connection, and should
therefore be the one which closes the connection for idle timeouts.

The idle_timeout, then, is here only to prevent bad clients from
leaving sockets open for a long time.

However, leaving idle connections open for a long time means that
the server has a large number of sockets used, where those sockets
aren't doing anything.  This can have other impacts.