reset the input pointer if it changes.  Fixex #5462
There are larger underlying problems, though.  The API to
fr_sbuff_terminal_search() is wrong.  It accepts an input pointer
"p", which points to somewhere in the middle of the sbuff.

As a result, if the sbuff is shifted due to reads, the sbuff->p
pointer changes, and the "cached" p passed to fr_sbuff_terminal_search()
points to somewhere wild.

As a result, most _callers_ of fr_sbuff_terminal_search() are
likely wrong, too, as they cache p = sbuff->p, and don't expect
that pointer to change under them.