Close path opened by fr_mkdir() to avoid fd leak (CID #1636671)
In fr_bio_fd_reopen(), if cfg->mkdir is true, it tries to create
a directory and returns -1 if it fails. If it succeeds, the
path number of the created direcctory is stored in fd...and the
next use of fd is assigning the result of open() to it, leaving
the directory open and accumulating inaccessible open files.