use "mode" instead of replicate / synchronous / originate
as it is more descriptive.  The other configuration flags also
can be configured in contradictory or conflicting ways.

Perhaps we could add a non-synchronous proxy, but that is likely
a good idea only if there's no Proxy-State in the packet.
i.e. we received the packet from a NAS, which has it's own retransmission
timers, and those timers are almost always garbage.