better handle "queue full" states
we already have a request state which indicates that the request
is owned by the queue.  We add a new request state which says that
the queue code should free the request on dequeue.

Double-checked the rest of the code, and added more cleanups
for requests which should handle the "queue full" case a bit
better.