Without &, bare words are more often attribute references.
Add a flag to the call_env API which says that this thing might
be an attribute reference, but only if it's a bare word.  And it's
not _required_ to be an attribute reference.

Update rlm_linelog to use the new flag.

Now that we can't key off of '&' to say "bare word is an attribute",
we must instead use a slightly more complex method.