clean up attribute parsing.
a conf item can be required to be an attribute.  However, right
now only "sometimes" and "sqlcounter" do this.  That needs to be
fixed.

If it is marked up as an attribute, then call the attribute parser,
and not the generic tmpl parser which then tries everything.

Remove cast of attribute.  The rest of the parser wouldn't allow it,
so that code was never used.