check for more corner cases on decoding
the tag should be what we expect, or at least be compatible with
what we expect.

On decoding set / sequence, ensure that we mark all input data
as consumed, even if we run out of children to decode.  That
way the decoder can move to the next DER item, instead of starting
to decode the next item from inside of the set / sequence.

In other words, we don't want the decoder wandering up and down
the eastern seaboard for three days, looking for Vancouver.