set tag for :V properly.
Hoist setting of new->tag to before we call fr_pair_value_from_str()

when converting LITERAL to DATA in modcall_fixup_update(), check
for TAG_VALUE, and then parse the tag from the value.  Once that's
done, parse the value with from the remaining part of the string.

Add tests