update the tmpl_rules for argument parsing
so that we don't do casts, etc. of the function arguments

add more checks, and produce better errors

when checking function arguments, we no longer assume that all
of the arguments have been parsed as a string.  Instead, they
may be parsed as some other data type.  We can then cast the
input expression to the expected data type for the argument.

this change also changes the expected output, where functions like
%md5() now print their parsed arguments as octet strings.