Py_GetPath is deprecated and will be remove.
People can just print sys.path if they need to see what it is

It would be nice for the python API documentation to help here.
It says Py_GetPath is deprecated, and to look at sys.path instead.
But it doesn't say how to get sys.path from the API.  A quick look
around shows how to get sys.path from the Python code, but not
from the API.

Instead of doing the right thing, we just remove the deprecated
code.