move to new functionality, mostly.
and go back to "deleting edit things is an ignore, not an error"
too many tests needed to change, and it doesn't seem to be worth
it.

Some keyword tests haven't been updated.  It's currently not
possible to loop over children of a list, when the children are of
different types.  Perhaps the solution there could be to use "void"
for the type?  So it doesn't create a local attribute for the
value.

the nested foreach doesn't work, as the nested local variable doesn't
change value.  Perhaps it isn't being deleted when the inner loop
exits