don't assume that the xlat outputs are string
they could be something else.  We never noticed before because
the rest of the code just mashed all of the switch keys to string,
too.

Now that we're a bit smarter, we can automatically determine the
data types for xlat functions.  So the key expansion for switch
statements should always cast to value_box, and not to string